# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ed4918b8a93b038ff5788e8
  title: My_team
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5ed4918b8a93b038ff5788fa
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Number
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5ed4918b8a93b038ff5788f4
      summary: Change password of user from admin
      tags:
        - User
      parameters:
        - name: My_team
          x-skaffolder-type: Decimal
          in: path
          description: My_team
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5ed4918b8a93b038ff5788f9
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles:
        - 5ed4918b8a93b038ff5788f4
      summary: CRUD ACTION create
      tags:
        - User
      parameters:
        - name: My_team
          x-skaffolder-type: Decimal
          in: path
          description: My_team
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5ed4918b8a93b038ff5788fc
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5ed4918b8a93b038ff5788f5
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ed4918b8a93b038ff5788fb
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5ed4918b8a93b038ff5788f7
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ed4918b8a93b038ff5788f2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    My_team:
      x-skaffolder-id: 5ed492b78a93b038ff5789d1
      x-skaffolder-id-db: 5ed4918b8a93b038ff5788e9
      x-skaffolder-id-entity: 5ed492b78a93b038ff5789cf
      x-skaffolder-url: /my_team
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Ojas:
          type: object
          x-skaffolder-id-attr: 5ed492b78a93b038ff5789d0
          x-skaffolder-type: Custom
      x-skaffolder-relations: 
      required:
        - _id
    User:
      x-skaffolder-id: 5ed4918b8a93b038ff5788f2
      x-skaffolder-id-db: 5ed4918b8a93b038ff5788e9
      x-skaffolder-id-entity: 5ed4918b8a93b038ff5788ea
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788ed
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788ee
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788ec
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788f0
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788ef
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5ed4918b8a93b038ff5788eb
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
  x-skaffolder-db:
    - x-skaffolder-id: 5ed4918b8a93b038ff5788e9
      x-skaffolder-name: My_team_db
  x-skaffolder-page:
    - x-skaffolder-id: 5ed4918b8a93b038ff5788f1
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5ed4941b8a93b038ff5789d7
      x-skaffolder-name: My team
      x-skaffolder-url: https//zoomus/
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
  x-skaffolder-roles: 
    - x-skaffolder-id: 5ed4918b8a93b038ff5788f4
      x-skaffolder-name: ADMIN